Main

Description

This module is an extensively documented walkthrough for typechecking a basic functional language using the Hindley-Damas-Milner algorithm.

It can be used in three different forms:

The source is written in literate programming style, so you can almost read it from top to bottom, minus some few references to later.

Runnable in GHCi.

The Haddock output yields a nice overview over the definitions given. It's not as good of a read as the source since many of the important inter-code comments are not visible.

Preliminaries

Prettyprinting

class Pretty a where

A prettyprinter class. Similar to Show, but with a focus on having human-readable output as opposed to being valid Haskell.

Methods

ppr :: a -> Text

Names

newtype Name

A name is an identifier in the language we're going to typecheck. Variables on both the term and type level have Names, for example.

Constructors

Name Text

Monotypes

data MType

A monotype is an unquantified/unparametric type. Monotypes are the inner building blocks of all types. Examples of monotypes are Int, a, a -> b.

In formal notation, MTypes are often called τ (tau) types.

Constructors

TVar Name
a

TFun MType MType
a -> b

TConst Name
Int, (), …

TList MType
[a]

TEither MType MType
Either a b

TTuple MType MType
(a,b)

freeMType :: MType -> Set Name

The free variables of an MType. This is simply the collection of all the individual type variables occurring inside of it.

Example: The free variables of a -> b are a and b.

substMType :: Subst -> MType -> MType

Apply a substitution to all known variables contained in an MType. Variables not mentioned are left unchanged.

This operation will not change the MType constructors that may be contained inside the MType recursively, it only acts on type variables.

Polytypes

data PType

A polytype is a monotype universally quantified over a number of type variables. In Haskell, all definitions have polytypes, but since the forall is implicit they look a bit like monotypes, maybe confusingly so. For example, the type of "1 :: Int" is actually "forall nothing. Int", and the type of "id" is "forall a. a -> a", although GHC displays it as "a -> a".

A polytype claims to work "for all imaginable type parameters", very similar to how a lambda claims to work "for all imaginable value parameters". We can insert a value into a lambda's parameter to evaluate it to a new value, and similarly we'll later insert types into a polytype's quantified variables to gain new types.

Example: in a definition id :: forall a. a -> a, the a after the ∀ ("forall") is the collection of type variables, and a -> a is the MType quantified over. When we have such an id, we also have its specialized version Int -> Int available. This process will be the topic of the type inference/unification algorithms.

In formal notation, PTypes are often called σ (sigma) types.

Constructors

Forall (Set Name) MType

freePType :: PType -> Set Name

The free variables of a PType are the free variables of the contained MType, except those universally quantified.

Example: foo :: forall a. a -> b -> a would be a PType in which b is a free type variable, while a is bound (via the forall).

substPType :: Subst -> PType -> PType

Apply a substitution to a PType, replacing all known variables in the contained MType except the ones universally quantified.

Invariant: the quantified variables are not changed by the operation.

The environment

newtype Env

The environment consists of all the values available in scope, and their associated polytypes.

Conceptually, the environment also contains all the things you can form by combining its elements, but this behaviour is modeled by the inference rules and not by the environment itself. This is a good thing, since the environment makes infinitely many things available to us: when we have z, then we also have y -> z and x y -> z and so on available to us.

In Haskell terms, the environment consists of all the things you currently have available. So if you import the Prelude, your environment consists of

id        →  ∀a. a→a
map       →  ∀a b. (a→b) → [a] → [b]
putStrLn  →  ∀∅. String → IO ()
…

Constructors

Env (Map Name PType)

freeEnv :: Env -> Set Name

The free variables of an Environment are all the free variables of the PTypes it contains.

substEnv :: Subst -> Env -> Env

Performing a Substitution in an Environment means performing that substituion on all the contained PTypes.

Substitutions

newtype Subst

A substitution is a mapping from type variables to MTypes. Applying a substitution means applying those replacements.

If you want to unify Int -> Int with a -> a, after inferring that a should be Int, that substituion has to be performed on the latter value to get rid of all the as.

A maybe more intuitive way that will come in handy for compose is to view a Substitution as a chain of

Constructors

Subst (Map Name MType)

empty :: Subst

The empty substituion holds nothing, and is the identity of compose.

compose :: Subst -> Subst -> Subst

Combine two substitutions. Applying the resulting substitution means applying the right hand side substitution first, and then the left hand side.

In the implementation of compose, we don't apply the substitution to anything of course, as we do not yet have anything to apply it to available. Because of this, we have to incorporate the first (higher priority) substitution substitution into the second one beforehand.

Typechecking

Typechecking does two things:

If two types are not immediately identical, attempt to unify them to get a type compatible with both of them

infer the most general type of a value by comparing the values in its definition with the Environment

Inference context

newtype Infer a

The inference type holds a supply of unique names, and can fail with a descriptive error if something goes wrong.

Constructors

Infer (ExceptT InferError (State [Text]) a)

data InferError

Constructors

CannotUnify MType MType
Two types that don't match were attempted to be unified.

OccursCheckFailed Name MType
A TVar is bound to an MType that already contains it.

The canonical example of this is λx. x x, where the first x in the body has to have type a → b, and the second one a. Since they're both the same x, this requires unification of a with a → b, which only works if a = a → b = (a → b) → b = …, yielding an infinite type.

UnknownIdentifier Name
The value of an unknown identifier was read.

OutOfFreshNames
The supply of fresh variable names has run out.

runInfer

Arguments

:: [Text]
Supply of variable names.

-> Infer a
Inference data

-> Either InferError a

Evaluate a value in an Inference context.

throw :: InferError -> Infer a

Throw an InferError in an Inference context.

Unification

Unification describes the process of making two different types compatible by specializing them where needed. A desirable property to have here is being able to find the most general unifier. Luckily, we'll be able to do that in our type system.

unify :: (MType, MType) -> Infer Subst

The unification of two MTypes is the most general substituion that can be applied to both of them in order to yield the same result.

Example: trying to unify a -> b with c -> (Int -> Bool) will result in a substitution of a for c, and b for Int -> Bool.

bindVariableTo :: Name -> MType -> Infer Subst

Build a Substitution that binds a Name of a TVar to an MType. The resulting substitution should be idempotent, i.e. applying it more than once to something should not be any different from applying it only once.

In the simplest case, this just means building a substitution that just does that.

Substituting a Name with a TVar with the same name unifies a type variable with itself, and the resulting substitution does nothing new.

If the Name we're trying to bind to an MType already occurs in that MType, the resulting substitution would not be idempotent: the MType would be replaced again, yielding a different result. This is known as the Occurs Check.

Type inference

Type inference is the act of finding out a value's type by looking at the environment it is in, in order to make it compatible with it.

In literature, the Hindley-Damas-Milner inference algorithm ("Algorithm W") is often presented in the style of logical formulas, and below you'll find that version along with code that actually does what they say.

These formulas look a bit like fractions, where the "numerator" is a collection of premises, and the denominator is the consequence if all of them hold.

Example:

Γ ⊢ even : Int → Bool   Γ ⊢ 1 : Int
-----------------------------------
         Γ ⊢ even 1 : Bool

means that if we have a value of type Int → Bool called "even" and a value of type Int called 1, then we also have a value of type Bool via even 1 available to us.

The actual inference rules are polymorphic versions of this example, and the code comments will explain each step in detail.

The language: typed lambda calculus

data Exp

The syntax tree of the language we'd like to typecheck. You can view it as a close relative to simply typed lambda calculus, having only the most necessary syntax elements.

Example: the term let y = λx. f x in z would be represented by

ELet "y"
     (EAbs "x"
           (EApp (EVar "f")
                 (EVar "x")))
     (EVar z)

Constructors

ELit Lit
True, 1

EVar Name
x

EApp Exp Exp
f x

EAbs Name Exp
x -> e

ELet Name Exp Exp
let x = e in e'

data Lit

Literals we'd like to support. Since we can't define new data types in our simple type system, we'll have to hard-code the possible ones here.

Constructors

LBool Bool

LInteger Integer

Some useful definitions

fresh :: Infer MType

Generate a fresh Name in a type Inference context. An example use case of this is η expansion, which transforms f into λx. f x, where "x" is a new name, i.e. unbound in the current context.

liftFresh :: MType -> PType

Lift a freshly generated MType into a PType by quantifying over no variables. This is only safe to use if no variable duplication can be ensured, as is the case when the argument is fresh. Violating this rule will result in (possibly silent!) name clashes.

Example:

a → b  ⇒  ∀∅ a → b

extendEnv :: Env -> (Name, PType) -> Env

Add a new binding to the environment.

The Haskell equivalent would be defining a new value, for example in module scope or in a let block. This corresponds to the "comma" operation used in formal notation,

Γ, x:σ  ≡  extendEnv Γ (x,σ)

Inferring the types of all language constructs

infer :: Env -> Exp -> Infer (Subst, MType)

Infer the type of an Expression in an Environment, resulting in the Exps MTypealong with a substitution that has to be done in order to reach this goal.

This is widely known as Algorithm W.

inferLit :: Lit -> Infer (Subst, MType)

Literals such as True and '1' have their types hard-coded.

inferVar :: Env -> Name -> Infer (Subst, MType)

Inferring the type of a variable is done via

x:σ ∈ Γ   τ = instantiate(σ)
----------------------------  [Var]
           Γ ⊢ x:τ

which simply means that if x is available in polymorphic form, then we have it available in all possible instantiations of that σ type. This allows us to take a σ type, and specialize it to our specific needs.

lookupEnv :: Env -> Name -> Infer PType

Look up the PType of a Name in the Environment.

To give a Haskell analogon, looking up id when Prelude is loaded, the resulting PType would be id's type, namely forall a. a -> a.

instantiate :: PType -> Infer MType

Bind all quantified variables of a PType to fresh type variables.

Example: instantiating forall a. a -> b -> a results in the MType a -> b -> a, where a is a fresh name (to avoid shadowing issues).

You can picture the PType to be the prototype converted to an instantiated MType, which can now be used in the unification process.

Another way of looking at it is by simply forgetting which variables were quantified over, carefully avoiding name clashes when doing so.

inferApp :: Env -> Exp -> Exp -> Infer (Subst, MType)

Function application captures the fact that if we have a function and an argument we can give to that function, we also have the result value of the result type available to us.

Γ ⊢ f : fτ   Γ ⊢ x : xτ   fxτ = fresh   unify(fτ, xτ → fxτ)
-----------------------------------------------------------  [App]
                      Γ ⊢ f x : fxτ

This rule is however a bit backwards to our normal way of thinking about typing function application: instead of applying a function to a value and investigating the result, we hypothesize the function type xτ → fxτ of mappting the argument xτ to the result type fxτ, and make sure this mapping unifies with the function f:fτ given.

inferAbs :: Env -> Name -> Exp -> Infer (Subst, MType)

Lambda abstraction is based on the fact that when we introduce a new variable, the resulting lambda maps from that variable's type to the type of the body.

τ = fresh   σ = liftFresh(τ)   Γ, x:σ ⊢ e:τ'
--------------------------------------------  [Abs]
               Γ ⊢ λx.e : τ→τ'

Here, Γ, x:τ is Γ extended by one additional mapping, namely x:τ.

Abstraction is typed by extending the environment by a new MType, and if under this assumption we can construct a function mapping to a value of that type, we can say that the lambda takes a value and maps to it.

inferLet :: Env -> Name -> Exp -> Exp -> Infer (Subst, MType)

A let binding allows extending the environment with new bindings in a principled manner. To do this, we first have to typecheck the expression to be introduced. The result of this is then generalized to a PType, since let bindings introduce new polymorphic values, and then added to the environment. Now we can finally typecheck the body of the "in" part of the let binding.

Γ ⊢ e:τ   σ = gen(Γ,τ)   Γ, x:σ ⊢ e':τ'
---------------------------------------  [Let]
        Γ ⊢ let x = e in e' : τ'

generalize :: Env -> MType -> PType

Generalize an MType to a PType by universally quantifying over all the type variables contained in it, except those already mentioned in the environment.

Example: Generalizing forall a. a -> b -> a yields forall a b. a -> b -> a.

gen(Γ,τ) = ∀{α}. σ
    where {α} = free(τ) – free(Γ)

Testing

A small custom Prelude

prelude :: Env

(~>) :: MType -> MType -> MType infixr 9

Synonym for TFun to make writing type signatures easier.

Instead of

Forall ["a","b"] (TFun "a" (TFun "b" "a"))

we can write

Forall ["a","b"] ("a" ~> "b" ~> "a")

defaultSupply :: [Text]

Supply to draw fresh type variable names from

Run it!

main :: IO ()

Run type inference on a cuple of values

lambda :: [Name] -> Exp -> Exp

Build multiple lambda bindings.

Instead of

EAbs "f" (EAbs "x" (EApp "f" "x"))

we can write

lambda ["f", "x"] (EApp "f" "x")

for

λf x. f x

apply :: Exp -> [Exp] -> Exp

Apply a function to multiple arguments.

Instead of

EApp (EApp (EApp "f" "x") "y") "z")

we can write

apply "f" ["x", "y", "z"]

for

f x y z

int :: Integer -> Exp

Construct an integer literal.

bool :: Bool -> Exp

Construct a boolean literal.

showType

Arguments

:: Env
Starting environment, e.g. prelude.

-> [Text]
Fresh variable name supply. Should be non-empty.

-> Exp
Expression to typecheck

-> Text
Text representation of the result. Contains an error message on failure.

Convenience function to run type inference algorithm